Update README.md
This is a report/work-in-progress that began a few years ago.  After collecting all of the EDL data for 5 years or so on each vessel, and wanting to address the top 20% of the equipment causing 80% of our downtime that was found in the Pareto Analysis, resulted in Deep Dives for all of the vessels.  Many of the cross-functional teams found multiple ways of addressing the following questions:
1.       Are we using the System correctly (within specified OEM parameters and environmental conditions)?
2.       Is the System engineered for how the vessel requires the System to function?
3.       Do we have to right parts in place for use, and lead time to acquire spares for future failure(s)?
4.       Does the Standard Job adequately address the failure(s) seen? 
So in order to answer them, I created and compiled all of the data required from multiple sources and when it came time to meet with our SME’s and cross-functional team, this allowed the data to be more convenient and easily navigated throughout the size and complexity of the meetings, with my team leading each of these for final recommendations and actions required to fix the vessel issues.